save idl1 files

+++ parallel_worker.py
@@ -78,6 +78,9 @@ def setup_multi_job(setup, job):
     else:
         print("write_ts flag unrecognised, stopped")
         exit(1)
+    if setup.save_idl1 == 1:
+        mkdir(setup.common_wd + "/idl1_folder/")
+        job.output.update({'idl1_folder' : setup.common_wd + "/idl1_folder"})
     return
 
 
@@ -159,7 +162,11 @@ def run_multi( job, atom, atmos):
 
             fbin.close()
             faux.close()
-        os.remove('./IDL1')
+        if setup.save_idl1 == 0:
+            os.remove('./IDL1')
+        elif setup.save_idl1 == 1:
+            destin = job.output[idl1_folder] + "/idl1.%s_%s_A(X)%5.5" %(atmos.id, atom.element, atom.abund)
+            shutil.mv('./IDL1', destin)
     # no IDL1 file created after the run
     else:
         print("IDL1 file not found for %s A(X)=%.2f" %(atmos.id, atom.abund))
